api/requirements-vercel.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
faiss-cpu==1.7.4
rank-bm25==0.2.2
scikit-learn==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
